Move RenderMethod to helper
