+++ helper/render_method.go
@@ -1,4 +1,4 @@
-package model
+package helper
 
 import (
 	"fmt"
