+++ wce/wce_wld.go
@@ -7,7 +7,6 @@ import (
 	"strings"
 
 	"github.com/xackery/quail/helper"
-	"github.com/xackery/quail/model"
 	"github.com/xackery/quail/raw"
 	"github.com/xackery/quail/raw/rawfrag"
 )
@@ -1975,7 +1974,7 @@ func (e *MaterialDef) ToRaw(wce *Wce, rawWld *raw.Wld) (int16, error) {
 	}
 
 	wfMaterialDef := &rawfrag.WldFragMaterialDef{
-		RenderMethod:  model.RenderMethodInt(e.RenderMethod),
+		RenderMethod:  helper.RenderMethodInt(e.RenderMethod),
 		RGBPen:        e.RGBPen,
 		Brightness:    e.Brightness,
 		ScaledAmbient: e.ScaledAmbient,
@@ -2055,7 +2054,7 @@ func (e *MaterialDef) FromRaw(wce *Wce, rawWld *raw.Wld, frag *rawfrag.WldFragMa
 		e.SimpleSpriteTag = rawWld.Name(spriteDef.NameRef())
 	}
 	e.Tag = rawWld.Name(frag.NameRef())
-	e.RenderMethod = model.RenderMethodStr(frag.RenderMethod)
+	e.RenderMethod = helper.RenderMethodStr(frag.RenderMethod)
 	e.RGBPen = frag.RGBPen
 	e.Brightness = frag.Brightness
 	e.ScaledAmbient = frag.ScaledAmbient
@@ -4216,7 +4215,7 @@ func (e *Sprite3DDef) ToRaw(wce *Wce, rawWld *raw.Wld) (int16, error) {
 				BackTree:      node.BackTree,
 				VertexIndexes: node.Vertices,
 
-				RenderMethod: model.RenderMethodInt(node.RenderMethod),
+				RenderMethod: helper.RenderMethodInt(node.RenderMethod),
 			}
 
 			if node.Pen.Valid {
@@ -4297,7 +4296,7 @@ func (e *Sprite3DDef) FromRaw(wce *Wce, rawWld *raw.Wld, frag *rawfrag.WldFragSp
 			FrontTree:    bspNode.FrontTree,
 			BackTree:     bspNode.BackTree,
 			Vertices:     bspNode.VertexIndexes,
-			RenderMethod: model.RenderMethodStr(bspNode.RenderMethod),
+			RenderMethod: helper.RenderMethodStr(bspNode.RenderMethod),
 		}
 
 		if bspNode.RenderFlags&0x01 == 0x01 {
@@ -8078,7 +8077,7 @@ func (e *Sprite2DDef) ToRaw(wce *Wce, rawWld *raw.Wld) (int16, error) {
 	}
 	wfSprite2D := &rawfrag.WldFragSprite2DDef{
 		Scale:        e.Scale,
-		RenderMethod: model.RenderMethodInt(e.RenderMethod),
+		RenderMethod: helper.RenderMethodInt(e.RenderMethod),
 	}
 
 	if e.DepthScale.Valid {
@@ -8280,7 +8279,7 @@ func (e *Sprite2DDef) FromRaw(wce *Wce, rawWld *raw.Wld, frag *rawfrag.WldFragSp
 		e.Pitches = append(e.Pitches, pitch)
 	}
 
-	e.RenderMethod = model.RenderMethodStr(frag.RenderMethod)
+	e.RenderMethod = helper.RenderMethodStr(frag.RenderMethod)
 	if frag.RenderFlags&0x01 == 0x01 {
 		e.Pen.Valid = true
 		e.Pen.Uint32 = frag.RenderPen
